Update to OTP File Encryption for cross-compatibility with existing SDKs. Changes to AES256-CTR to match existing SDKs
